change debug flag to print info before solving
Before, the debug flag ran the solve function on a test board. This was
more useful before the -f flag was implemented, but now it has been virtually
replaced by that flag. I decided that it would be much more useful to see information
like what the board was simplified to and the value of nums instead,
because that functionality was prevously not available without
recompiling.