move definition of variables to fix error
Compiling on gcc 10.1.0-2 produced a multiple definition error due to
boxes and unsolvedBoard being in window.h which was included multiple
times. The definition of these variables has been moved to window.c
from window.h because they do not need to be visible from main.c, and
this fixed the issue as well.